% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{create_biom}
\alias{create_biom}
\title{Create BIOM object from raw HDF5 array}
\usage{
create_biom(h5array)
}
\arguments{
\item{h5array}{A raw HDF5 array read from a BIOM file.}
}
\value{
A [biomformat::biom()] object.
}
\description{
`create_biom()` is an internal function used by [load_biom()], that more or
less replicates [biomformat::read_hdf5_biom()].
}
