% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{create_row_data}
\alias{create_row_data}
\title{Recreate observation metadata}
\usage{
create_row_data(biom, ranks)
}
\arguments{
\item{biom}{A BIOM object.}

\item{ranks}{A character vector of taxonomic ranks in order.}
}
\value{
A [data.frame()] with observation metadata.
}
\description{
`create_row_data()` is an internal function used by [load_biom()], that
returns a copy of the BIOM object's observation metadata, where the headers
of the taxonomy columns have been replaced with the correct taxonomic rank
names.
}
