% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{get_ranks}
\alias{get_ranks}
\title{Get taxonomic ranks from observation group metadata}
\usage{
get_ranks(h5array)
}
\arguments{
\item{h5array}{A raw HDF5 array read from a BIOM file.}
}
\value{
A character vector of taxonomic ranks in order.
}
\description{
`get_ranks()` is an internal function used by [load_biom()], that reads the
string of semi-colon separated ranks from the group observation metadata.
}
