% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_biom}
\alias{load_biom}
\title{Load BIOM format file to `TreeSummarizedExperiment`}
\usage{
load_biom(filename)
}
\arguments{
\item{filename}{A character vector with one element that denotes the file
path to a BIOM file.}
}
\value{
A [TreeSummarizedExperiment::TreeSummarizedExperiment()] object.
}
\description{
Load BIOM format file to `TreeSummarizedExperiment`
}
\examples{
tse <- load_biom("profiles.biom")
}
